fixed all issue with items
There was some issues with items and network. For instance items would not show to be equipped on the client when the host can see it. Basically fixed all item issue with networking and item authority situation . Also made some changes to the format of the weapon and armor class.